fixes karemis hazelcast api
