+++ hazelcast-extension/src/main/java/org/jnosql/artemis/hazelcast/keyvalue/HazelcastRepository.java
@@ -12,8 +12,16 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.hazelcast.kv;
+package org.jnosql.artemis.hazelcast.keyvalue;
 
 
-public interface PersonRepository extends HazelcastRepository<Person, String> {
+import jakarta.nosql.mapping.Repository;
+
+/**
+ * The hazelcat {@link Repository}
+ *
+ * @param <T> the entity type
+ * @param <K> the id entity type
+ */
+public interface HazelcastRepository<T, K> extends Repository<T, K> {
 }
