+++ hazelcast-extension/src/main/java/org/jnosql/artemis/hazelcast/keyvalue/HazelcastTemplate.java
@@ -12,10 +12,10 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.hazelcast.kv;
+package org.jnosql.artemis.hazelcast.keyvalue;
 
 import com.hazelcast.query.Predicate;
-import jakarta.nosql.mapping.kv.KeyValueTemplate;
+import jakarta.nosql.mapping.keyvalue.KeyValueTemplate;
 
 import java.util.Collection;
 import java.util.Map;
