+++ hazelcast-extension/src/main/java/org/jnosql/artemis/hazelcast/keyvalue/ParamUtil.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.hazelcast.kv;
+package org.jnosql.artemis.hazelcast.keyvalue;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
