+++ hazelcast-extension/src/test/java/org/jnosql/artemis/hazelcast/keyvalue/HazelcastRepositoryProxyTest.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.hazelcast.kv;
+package org.jnosql.artemis.hazelcast.keyvalue;
 
 import com.hazelcast.query.Predicate;
 import jakarta.nosql.mapping.Repository;
