+++ hazelcast-extension/src/test/java/org/jnosql/artemis/hazelcast/keyvalue/Person.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.hazelcast.kv;
+package org.jnosql.artemis.hazelcast.keyvalue;
 
 
 import jakarta.nosql.mapping.Column;
